fix: refaor the code
