+++ apps/web/components/Feature/Dashboard/ClientDashboardCreators.tsx
@@ -17,7 +17,7 @@ import { useTranslation } from "react-i18next";
 import UsersContent from "../Users/UsersContent";
 import { useLogout } from "@/hooks/auth/useLogout";
 import { useAuthSession } from "@/hooks/auth/useAuthSession";
-import { useCreatorDashboardProfileSync } from "@/hooks/auth/creatorProfileApi";
+import { useProfileSync } from "@/hooks/auth/useProfileSync";
 
 const ROUTABLE_DASHBOARD_VIEWS = new Set<string>([
   CREATORS_LABELS.OVERVIEW,
@@ -38,7 +38,7 @@ export default function ClientDashboardCreators() {
   const searchParams = useSearchParams();
   const { logout } = useLogout();
   const { getUser } = useAuthSession();
-  useCreatorDashboardProfileSync();
+  useProfileSync();
 
   const toggleSidebar = useCallback(() => {
     setOpen((p) => !p);
