+++ apps/web/hooks/auth/creatorProfileApi.ts
@@ -1,27 +1,19 @@
-"use client";
-
-import { useEffect } from "react";
-import { mergeStoredLoginUser } from "@/hooks/auth/useLogin";
-import { API } from "@/lib/http/api/endpoints";
-import { useGetAPI } from "@/lib/http/api/getApi";
 import {
-  displayCreatorName,
-  getAvatarUrl,
   parseCreatorNameFromFullName,
-  toOptionalString,
+  type CreatorProfilePatchBody,
   type ProfileForm,
 } from "@/utils/creatorProfile";
-import { PROFILE_FIELD_MAP } from "@/utils/profileFieldMap";
-
-const trim = (value: string) => value.trim();
 
 export {
+  applyCreatorProfileResponseToForm,
+  buildCreatorProfilePatchBody,
   displayCreatorName,
   EMPTY_CREATOR_BOOT,
   EMPTY_CREATOR_PROFILE_FORM,
   getAvatarUrl,
   readCreatorBoot,
   toOptionalString,
+  type CreatorProfilePatchBody,
 } from "@/utils/creatorProfile";
 
 export type CreatorProfileUser = {
@@ -58,24 +50,15 @@ export type GetCreatorProfileResponse = {
   data?: CreatorProfileApiData;
 };
 
-export type UpdateCreatorProfileBody = {
-  firstName?: string;
-  lastName?: string;
-  avatarUrl?: string | null;
-  companyName?: string;
-  phone?: string;
-  cvr?: string;
-  address?: string;
-  city?: string;
-  postalCode?: string;
-  regNumber?: string;
-  accountNumber?: string;
-};
+export type {
+  CreatorProfilePatchBody,
+  CreatorProfilePatchBody as UpdateCreatorProfileBody,
+} from "@/utils/creatorProfile";
 
 export type UpdateCreatorProfileResponse = {
   success?: boolean;
   message?: string;
-  data?: UpdateCreatorProfileBody;
+  data?: CreatorProfilePatchBody;
 };
 
 const str = (value: unknown) => (typeof value === "string" ? value.trim() : "");
@@ -109,75 +92,4 @@ export function mapCreatorProfileToForm(
     reg: str(bank?.registrationNumber),
     account: str(bank?.accountNumber),
   };
-}
-
-export function applyCreatorProfileResponseToForm(
-  form: ProfileForm,
-  data?: UpdateCreatorProfileBody,
-): ProfileForm {
-  const next = { ...form };
-
-  for (const [formKey, bodyKey] of PROFILE_FIELD_MAP) {
-    const value = data?.[bodyKey];
-    if (typeof value === "string") {
-      next[formKey] = value.trim();
-    }
-  }
-
-  return next;
-}
-
-export function buildCreatorProfilePatchBody(
-  form: ProfileForm,
-  saved: ProfileForm,
-  avatarDirty: boolean,
-  avatarImage: string | null,
-): UpdateCreatorProfileBody {
-  const body = PROFILE_FIELD_MAP.reduce<UpdateCreatorProfileBody>(
-    (acc, [formKey, bodyKey]) => {
-      const current = trim(form[formKey]);
-      const previous = trim(saved[formKey]);
-
-      if (current !== previous) {
-        acc[bodyKey] = current;
-      }
-
-      return acc;
-    },
-    {},
-  );
-
-  if (avatarDirty) {
-    body.avatarUrl = avatarImage ?? null;
-  }
-
-  return body;
-}
-
-export function useCreatorDashboardProfileSync() {
-  const profileQuery = useGetAPI<GetCreatorProfileResponse>(
-    API.auth.creatorProfile,
-    undefined,
-    {
-      retry: false,
-      refetchOnWindowFocus: false,
-    },
-  );
-
-  useEffect(() => {
-    const profile = profileQuery.data?.data;
-    if (!profile) return;
-
-    const form = mapCreatorProfileToForm(profile);
-
-    mergeStoredLoginUser({
-      email: toOptionalString(form.email),
-      fullName: toOptionalString(displayCreatorName(form)),
-      firstName: toOptionalString(form.firstName),
-      lastName: toOptionalString(form.lastName),
-      avatarUrl: getAvatarUrl(profile.user?.avatarUrl),
-    });
-  }, [profileQuery.data]);
-
-  return profileQuery;
 }
